+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 
+## [0.5.3](https://github.com/Blobfolio/argyle/releases/tag/v0.5.3) - 2022-03-29
+
+### New
+
+* `Argue::args_os`
+* `Argue::arg_os`
+* `Argue::first_arg_os`
+
+
+
 ## [0.5.2](https://github.com/Blobfolio/argyle/releases/tag/v0.5.2) - 2021-12-25
 
 ### New